+++ app/static/js_files/admin_reviews.js
@@ -648,12 +648,20 @@ function closeReviewModal() {
 
 // Fonction qui gère la suppression d'un avis
 async function deleteReview(id) {
+	const csrfToken = getCookie('csrf_access_token');
+    if (!csrfToken) {
+        console.error('Token CSRF manquant');
+        showFeedbackMessage('Session invalide, veuillez rafraichir la page', true);
+        return;
+    }
+
 	try {
 		const response = await fetch(`${API_REVIEWS_URL}/${id}`, {
 			method: 'DELETE',
 			credentials: 'include',
 			headers: {
-				'Content-Type': 'application/json'
+				'Content-Type': 'application/json',
+				'X-CSRF-TOKEN': csrfToken
 			}
 		});
